requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
structlog
python-jose[cryptography]
passlib[bcrypt]
requests
google-generativeai

# Test dependencies
pytest
httpx
pytest-mock
